It's working
You can hit enter to exit the save window.  You can hit escape to exit the help window.  I got rid of the menu panel on the bottom of the board.  I plan on adding a keybind to open a window to have a button menu instead of having it at the bottom of the board